requirements.txt
httpx~=0.17.0
nonebot2==2.0.0a11
nb-cli~=0.4.0
nonebot-test~=0.1.0
nonebot-plugin-apscheduler~=0.1.0
nonebot-adapter-cqhttp~=2.0.0a11.post2
